Update readme.md to render correctly on npmjs.com
My guess is that the display problems on npmjs.com are caused by the different indents for the start and end code-section markers.

npmjs.com uses a http://commonmark.org/ parser, which means it's not _exactly_ the same as github. This may just be that, or it may be a bug in the parser.